refactor : DAO에서 DB 연결 부분 분리
