+++ src/test/java/web/dao/BoardDaoTest.java
@@ -1,31 +1,21 @@
 package web.dao;
 
-import static org.junit.jupiter.api.Assertions.*;
-
-import java.sql.Connection;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import web.dto.ChessGameDto;
 
-public class BoardDaoTest {
+class BoardDaoTest {
 	private BoardDao boardDAO;
 
 	@BeforeEach
-	public void setup() {
+	void setup() {
 		boardDAO = new BoardDao();
 	}
 
 	@Test
-	public void connection() {
-		Connection con = boardDAO.getConnection();
-		assertNotNull(con);
-	}
-
-	@Test
-	public void addUser() throws Exception {
-		ChessGameDto chessGameDto = new ChessGameDto("5번", "ppppPPP", "white");
+	void addUser() throws Exception {
+		ChessGameDto chessGameDto = new ChessGameDto("7번", "ppppPPP", "white");
 		boardDAO.addBoard(chessGameDto);
 	}
 }
